requirements.txt
# Dependências principais
requests>=2.31.0
beautifulsoup4>=4.12.2
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0

# Dependências opcionais (não estritamente necessárias, mas úteis)
schedule>=1.2.1